+++ src/main/java/com/spring_jms/SpringJmsApplicatoin.java
@@ -5,6 +5,8 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import com.spring_jms.producer.MessageProducer;
+
 @SpringBootApplication
 public class SpringJmsApplicatoin implements CommandLineRunner {
